abstract embeddings no longer chunked
